Remove definitions from generated AST
This is a pretty fancy trick to implement a language with recursive
definitions using laziness. Usually the solution to implementing
recursive definitions would involve some lookup system to enable
"function calls", or using fixed-point combinators to get recursion.

However, since Haskell is lazy we can just inline recursive definitions
and we end up with a single term containing our whole program. We get
recursive definitions without adding any constructs! This means the
evaluator just sees terms and has no idea that they're even recursive.
Since the evaluator is also lazy, it only uses the terms it needs, so
evaluation is only as recursive as the program it's evaluating (as is
normal).

The one drawback here is that using this data structure greedily will
get you into trouble. For example, our pretty printer now hangs on
recursive definitions. For a normal person this is undesirable. Haskell
programmers are not normal however.